refactor: InputForm 공통 컴포넌트 추가 및 중복 컴포넌트 제거
- InputForm 공통 컴포넌트 추가
- SearchBar, GameInput 컴포넌트 제거
