+++ week3/assignment/src/components/views/GitHubSearch.jsx
@@ -1,4 +1,4 @@
-import SearchBar from '../github/SearchBar';
+import InputForm from '../common/InputForm';
 import UserCard from '../github/UserCard';
 import RecentSearches from '../github/RecentSearches';
 import useGitHubSearch from '../../hooks/useGitHubSearch';
@@ -9,7 +9,10 @@ function GitHubSearch() {
 
   return (
     <div className="p-4">
-      <SearchBar onSearch={getUserInfo} />
+      <InputForm
+        onSubmit={getUserInfo}
+        placeholder="Github 프로필을 검색해보세요."
+      />
       <RecentSearches 
         searches={recentSearches} 
         onRemove={removeSearch}
